Switch Menu to dot notation for subcomponents
