+++ packages/components/src/menu.link.test.basics.ts
@@ -1,35 +1,43 @@
 import { expect, fixture, html } from '@open-wc/testing';
-import CsMenuButton from './menu-button.js';
+import CsMenuLink from './menu.link.js';
 
-CsMenuButton.shadowRootOptions.mode = 'open';
+CsMenuLink.shadowRootOptions.mode = 'open';
 
 it('registers', async () => {
-  expect(window.customElements.get('cs-menu-button')).to.equal(CsMenuButton);
+  expect(window.customElements.get('cs-menu-link')).to.equal(CsMenuLink);
 });
 
 it('has defaults', async () => {
   // Required attributes are supplied and not asserted below. The idea is that
   // this test shouldn't fail to typecheck if these templates are eventually
   // typechecked, which means supplying required attributes up front.
-  const button = await fixture<CsMenuButton>(
-    html`<cs-menu-button label="Label"></cs-menu-button>`,
+  const link = await fixture<CsMenuLink>(
+    html`<cs-menu-link label="Label" url="/"></cs-menu-link>`,
   );
 
   // Not reflected. So no attribute assertions are necessary.
-  expect(button.privateActive).to.equal(false);
+  expect(link.privateActive).to.be.false;
 });
 
 it('can have a label', async () => {
-  const button = await fixture<CsMenuButton>(
-    html`<cs-menu-button label="Label"></cs-menu-button>`,
+  const link = await fixture<CsMenuLink>(
+    html`<cs-menu-link label="Label" url="/"></cs-menu-link>`,
   );
 
-  expect(button.shadowRoot?.textContent?.trim()).to.equal('Label');
+  expect(link.shadowRoot?.textContent?.trim()).to.equal('Label');
+});
+
+it('can have a URL', async () => {
+  const link = await fixture<CsMenuLink>(
+    html`<cs-menu-link label="Label" url="/"></cs-menu-link>`,
+  );
+
+  expect(link.url).to.equal('/');
 });
 
 it('can have an icon', async () => {
-  const button = await fixture<CsMenuButton>(
-    html`<cs-menu-button label="Label">
+  const link = await fixture<CsMenuLink>(
+    html`<cs-menu-link label="Label">
       <svg
         slot="icon"
         width="16"
@@ -47,10 +55,10 @@ it('can have an icon', async () => {
           d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L10.582 16.07a4.5 4.5 0 0 1-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 0 1 1.13-1.897l8.932-8.931Zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0 1 15.75 21H5.25A2.25 2.25 0 0 1 3 18.75V8.25A2.25 2.25 0 0 1 5.25 6H10"
         />
       </svg>
-    </cs-menu-button>`,
+    </cs-menu-link>`,
   );
 
-  const icon = button?.shadowRoot
+  const icon = link?.shadowRoot
     ?.querySelector<HTMLSlotElement>('slot[name="icon"]')
     ?.assignedElements()
     .at(0);
