+++ packages/components/src/menu.test.focus.ts
@@ -1,5 +1,5 @@
-import './menu-link.js';
 import './menu.js';
+import './menu.link.js';
 import { elementUpdated, expect, fixture, html } from '@open-wc/testing';
 import type CsMenu from './menu.js';
